Rename wasm_storage to wasm_mockups
