+++ clients/client-core/src/init/helpers.rs
@@ -9,7 +9,7 @@ use crate::{
 use config::NymConfig;
 use crypto::asymmetric::identity;
 #[cfg(target_arch = "wasm32")]
-use gateway_client::wasm_storage::SigningNyxdClient;
+use gateway_client::wasm_mockups::SigningNyxdClient;
 use gateway_client::GatewayClient;
 use gateway_requests::registration::handshake::SharedKeys;
 use rand::{seq::SliceRandom, thread_rng};
