+++ common/client-libs/gateway-client/src/lib.rs
@@ -17,7 +17,7 @@ pub mod error;
 pub mod packet_router;
 pub mod socket_state;
 #[cfg(target_arch = "wasm32")]
-pub mod wasm_storage;
+pub mod wasm_mockups;
 
 /// Helper method for reading from websocket stream. Helps to flatten the structure.
 pub(crate) fn cleanup_socket_message(
