+++ common/client-libs/gateway-client/src/wasm_mockups.rs
